+++ packages/esm-ugandaemr-app/src/dashboards/user/components/user-dashboard-header/user-dashboard-header.component.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Calendar, Location } from '@carbon/react/icons';
 import { formatDate, useSession } from '@openmrs/esm-framework';
-import ReportingHomeIllustration from './facility-dashboard-illustration.component';
-import styles from './facility-dashboard-header.scss';
+import ReportingHomeIllustration from './user-dashboard-illustration.component';
+import styles from './user-dashboard-header.scss';
 
-const FacilityDashboardHeader: React.FC<{ title?: string }> = ({ title }) => {
+const UserDashboardHeader: React.FC<{ title?: string }> = ({ title }) => {
   const { t } = useTranslation();
   const userSession = useSession();
   const userLocation = userSession?.sessionLocation?.display;
@@ -17,7 +17,7 @@ const FacilityDashboardHeader: React.FC<{ title?: string }> = ({ title }) => {
           <ReportingHomeIllustration />
           <div className={styles['page-labels']}>
             <p>{t('home', 'Home')}</p>
-            <p className={styles['page-name']}>{title ?? t('facility-dashboard', 'Facility Dashboard')}</p>
+            <p className={styles['page-name']}>{title ?? t('user-dashboard', 'User Dashboard')}</p>
           </div>
         </div>
         <div className={styles['right-justified-items']}>
@@ -34,4 +34,4 @@ const FacilityDashboardHeader: React.FC<{ title?: string }> = ({ title }) => {
   );
 };
 
-export default FacilityDashboardHeader;
+export default UserDashboardHeader;
