+++ packages/esm-ugandaemr-app/src/dashboards/user/components/user-dashboard-header/user-dashboard-header.scss
