+++ packages/esm-ugandaemr-app/src/dashboards/user/empty-state-illustration.component.tsx
